Move test code into project-point file
